go-to definition should return IDs instead of full ASTs. Also added loggers
