+++ presentation-compiler/src/test/scala/org/alephium/ralph/lsp/pc/search/gotodef/GoToAssignmentsInTxScriptSpec.scala
@@ -38,10 +38,10 @@ class GoToAssignmentsInTxScriptSpec extends AnyWordSpec with Matchers {
     "assigned variables exist" in {
       goToDefinition(
         """
-          |TxScript GoToAssignment(>>counter: U256<<) {
-          |  >>let mut counter = 0<<
+          |TxScript GoToAssignment(>>counter<<: U256) {
+          |  let mut >>counter<< = 0
           |  counte@@r = counter + 1
-          |  for (>>let mut counter = 0<<; counter <= 4; counter = counter + 1) {
+          |  for (let mut >>counter<< = 0; counter <= 4; counter = counter + 1) {
           |    counter = counter + 1
           |  }
           |}
