+++ presentation-compiler/src/test/scala/org/alephium/ralph/lsp/pc/search/gotodef/GoToBuiltInFunctionsSpec.scala
@@ -50,7 +50,7 @@ class GoToBuiltInFunctionsSpec extends AnyWordSpec with Matchers {
               |  }
               |}
               |""".stripMargin,
-          expected = Some("""fn assert!(condition:Bool, errorCode:U256) -> ()""")
+          expected = Some("""assert!""")
         )
       }
 
@@ -63,7 +63,7 @@ class GoToBuiltInFunctionsSpec extends AnyWordSpec with Matchers {
               |Interface TestBuiltIn {
               |  fn hello!() -> ()
               |
-              |  >>fn assert!() -> ()<<
+              |  fn >>assert!<<() -> ()
               |
               |  fn blah!() -> ()
               |}
@@ -92,7 +92,7 @@ class GoToBuiltInFunctionsSpec extends AnyWordSpec with Matchers {
             |  }
             |}
             |""".stripMargin,
-        expected = Some("""fn verifyAbsoluteLocktime!(lockUntil:U256) -> ()""")
+        expected = Some("""verifyAbsoluteLocktime!""")
       )
     }
 
@@ -111,7 +111,7 @@ class GoToBuiltInFunctionsSpec extends AnyWordSpec with Matchers {
             |  }
             |}
             |""".stripMargin,
-        expected = Some("""fn assert!(condition:Bool, errorCode:U256) -> ()""")
+        expected = Some("""assert!""")
       )
     }
   }
